# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jatek Food Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags:
        - menu
      summary: Get a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
      security:
        - bearerAuth: []
    patch:
      operationId: updateMenuItem
      tags:
        - menu
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
      security:
        - bearerAuth: []
      description: "RBAC: admin or owning restaurant_owner."
    delete:
      operationId: deleteMenuItem
      tags:
        - menu
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
      security:
        - bearerAuth: []
      description: "RBAC: admin or owning restaurant_owner."
...
  /restaurants/{restaurantId}/menu:
    get:
      operationId: listMenuItems
      tags:
        - menu
      summary: List menu items for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags:
        - menu
      summary: Create a menu item
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
      description: "RBAC: admin or owning restaurant_owner."
...
  /reviews:
    get:
      operationId: listReviews
      tags:
        - reviews
      summary: List reviews
      parameters:
        - name: restaurantId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
      security:
        - bearerAuth: []
    post:
      operationId: createReview
      tags:
        - reviews
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
      security:
        - bearerAuth: []
      description: JWT required.
  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags:
        - reviews
      summary: Delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
      security:
        - bearerAuth: []
      description: JWT required. Review author or admin only.
...
  /storage/objects/{path}:
    get:
      operationId: getObject
      tags:
        - storage
      summary: Serve an uploaded object
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object bytes
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to serve object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags:
        - storage
      summary: Request a presigned URL for file upload
      description: "RBAC: admin or restaurant_owner only."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to generate upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        name:
          type: string
        description:
          type:
            - string
            - "null"
        price:
          type: number
        imageUrl:
          type:
            - string
            - "null"
        category:
          type: string
        isAvailable:
          type: boolean
        isPopular:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - restaurantId
        - name
        - price
        - category
        - isAvailable
        - isPopular
        - createdAt
    CreateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        isAvailable:
          type: boolean
        isPopular:
          type: boolean
      required:
        - name
        - price
        - category
    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        isAvailable:
          type: boolean
        isPopular:
          type: boolean
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        restaurantId:
          type: integer
        orderId:
          type:
            - integer
            - "null"
        userName:
          type: string
        rating:
          type: integer
        comment:
          type:
            - string
            - "null"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - restaurantId
        - userName
        - rating
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        restaurantId:
          type: integer
        orderId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
      required:
        - restaurantId
        - rating
...
    UploadUrlRequest:
      type: object
      required:
        - name
        - size
        - contentType
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. image/jpeg).
    UploadUrlResponse:
      type: object
      required:
        - uploadURL
        - objectPath
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. /objects/uploads/uuid). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by /auth/login, /auth/verify-otp, or /backend/login. Some HTML document endpoints also accept the same token as a token query parameter.